--- a/bin/fuzz/fuzz_targets/protocol-block-announce-decode.rs
+++ b/bin/fuzz/fuzz_targets/protocol-block-announce-decode.rs
@@ -17,6 +17,6 @@
 
 #![no_main]
 
-libfuzzer_sys::fuzz_target!(|data: &[u8]| {
-    let _ = smoldot::network::protocol::decode_block_announce(data);
+libfuzzer_sys::fuzz_target!(|params: (&[u8], u8)| {
+    let _ = smoldot::network::protocol::decode_block_announce(params.0, usize::from(params.1) + 1);
 });