Close https://github.com/paritytech/smoldot/issues/2478

This is the last step of https://github.com/paritytech/smoldot/issues/2478: block headers.
I did block headers last because it's the largest change.