@tomaka 2022-07-28
The `header` benchmark:

Before:

```
Benchmarking headers/decode/325: Collecting 100 samples in estimated                                                                      headers/decode/325      time:   [34.944 ns 35.194 ns 35.487 ns]
                        thrpt:  [8.5294 GiB/s 8.6004 GiB/s 8.6619 GiB/s]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe
Benchmarking headers/decode_and_to_owned/325: Collecting 100 samples                                                                      headers/decode_and_to_owned/325                        
                        time:   [168.11 ns 168.79 ns 169.53 ns]
                        thrpt:  [1.7854 GiB/s 1.7932 GiB/s 1.8005 GiB/s]
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
Benchmarking headers/decode/16104: Collecting 100 samples in estimate                                                                     headers/decode/16104    time:   [226.22 ns 228.60 ns 231.42 ns]
                        thrpt:  [64.808 GiB/s 65.609 GiB/s 66.298 GiB/s]
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  6 (6.00%) high severe
Benchmarking headers/decode_and_to_owned/16104: Warming up for 3.0000                                                                     Benchmarking headers/decode_and_to_owned/16104: Collecting 100 sample                                                                     headers/decode_and_to_owned/16104                        
                        time:   [1.4537 µs 1.4719 µs 1.4989 µs]
                        thrpt:  [10.006 GiB/s 10.190 GiB/s 10.317 GiB/s]
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) high mild
  8 (8.00%) high severe
```

After:

```
Benchmarking headers/decode/325: Collecting 100 samples in estimated                                                                      headers/decode/325      time:   [32.944 ns 33.088 ns 33.242 ns]
                        thrpt:  [9.1054 GiB/s 9.1478 GiB/s 9.1877 GiB/s]
                 change:
                        time:   [-5.8124% -3.0929% +0.4026%] (p = 0.04 < 0.05)
                        thrpt:  [-0.4009% +3.1917% +6.1711%]
                        Change within noise threshold.
Found 10 outliers among 100 measurements (10.00%)
  10 (10.00%) high severe
Benchmarking headers/decode_and_to_owned/325: Collecting 100 samples                                                                      headers/decode_and_to_owned/325                        
                        time:   [153.75 ns 154.54 ns 155.49 ns]
                        thrpt:  [1.9467 GiB/s 1.9586 GiB/s 1.9687 GiB/s]
                 change:
                        time:   [-7.5653% -6.6383% -5.6039%] (p = 0.00 < 0.05)
                        thrpt:  [+5.9366% +7.1103% +8.1845%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild
Benchmarking headers/decode/16104: Collecting 100 samples in estimate                                                                     headers/decode/16104    time:   [215.99 ns 216.99 ns 218.15 ns]
                        thrpt:  [68.753 GiB/s 69.119 GiB/s 69.439 GiB/s]
                 change:
                        time:   [-4.8939% -3.4275% -1.6926%] (p = 0.00 < 0.05)
                        thrpt:  [+1.7217% +3.5491% +5.1457%]
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  7 (7.00%) high mild
  5 (5.00%) high severe
Benchmarking headers/decode_and_to_owned/16104: Warming up for 3.0000                                                                     Benchmarking headers/decode_and_to_owned/16104: Collecting 100 sample                                                                     headers/decode_and_to_owned/16104                        
                        time:   [1.3710 µs 1.3751 µs 1.3800 µs]
                        thrpt:  [10.868 GiB/s 10.907 GiB/s 10.940 GiB/s]
                 change:
                        time:   [-8.0238% -6.2985% -4.7116%] (p = 0.00 < 0.05)
                        thrpt:  [+4.9446% +6.7219% +8.7238%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
```

I don't understand how it can be faster, but won't complain :shrug: